Update GameCommands.cs
Text updates. I changed the text for "equip" command but I did not change the command itself. Would prefer "Handle" to equip. "In Hand" and "Handled" etc
Let's try a longer spawn time too - maybe 30s? I did not change this command, just wanted to mention it.
I notice the capture command was already changed to purchase in this build so no changes there but I did change some of the text as you'll see.